Switch from int to long, use invariant culture, use singleton codec

+++ test/JsonApiDotNetCoreTests/IntegrationTests/IdObfuscation/BankAccountsController.cs
@@ -5,5 +5,5 @@ using Microsoft.Extensions.Logging;
 namespace JsonApiDotNetCoreTests.IntegrationTests.IdObfuscation;
 
 public sealed class BankAccountsController(
-    IJsonApiOptions options, IResourceGraph resourceGraph, ILoggerFactory loggerFactory, IResourceService<BankAccount, int> resourceService)
+    IJsonApiOptions options, IResourceGraph resourceGraph, ILoggerFactory loggerFactory, IResourceService<BankAccount, long> resourceService)
     : ObfuscatedIdentifiableController<BankAccount>(options, resourceGraph, loggerFactory, resourceService);
